moved method belonging to services from the metamodel
