+++ fennec/apps/repository/utils.py
@@ -1,11 +1,6 @@
-from StringIO import StringIO
-
-from rest_framework.parsers import JSONParser
+from fennec.apps.metamodel.services import change_to_object
 
 from fennec.apps.metamodel.utils import ProjectBasicInfo, BranchBasicInfo, SandboxBasicInfo
-from fennec.apps.metamodel.serializers import SchemaSerializer, NamespaceSerializer, TableSerializer, ColumnSerializer, \
-    IndexSerializer, ForeignKeySerializer, LayerSerializer, TableElementSerializer, RelationshipElementSerializer, \
-    DiagramSerializer
 from fennec.apps.repository.models import Sandbox, Branch, BranchRevision, SandboxChange, \
     BranchRevisionChange
 from fennec.apps.metamodel.models import Change
@@ -13,37 +8,6 @@ from fennec.apps.metamodel.models import Change
 
 __author__ = 'Darko'
 
-
-def change_to_object(change):
-    stream = StringIO(change.content)
-    # print "Change content:"  + str(change.content)
-    data = {}
-    try:
-        data = JSONParser().parse(stream)
-    except Exception as e:
-        pass
-        # print e
-
-    serializer = switch_type(change.object_type)(data=data)
-    if not serializer.is_valid():
-        print serializer.errors
-
-    return serializer.object
-
-
-def switch_type(object_type):
-    return {
-        'Schema': SchemaSerializer,
-        'Namespace': NamespaceSerializer,
-        'Table': TableSerializer,
-        'Column': ColumnSerializer,
-        'Index': IndexSerializer,
-        'ForeignKey': ForeignKeySerializer,
-        'Diagram': DiagramSerializer,
-        'Layer': LayerSerializer,
-        'TableElement': TableElementSerializer,
-        'RelationshipElement': RelationshipElementSerializer
-    }.get(object_type, None)
 
 
 def branch_from_branch_revision(branch_rev, account, name, type=None,
